Get the extension board header into place :-)))
Boy, what a long journey to finally get here.

A dirty trick is involved, though. To allow a track between pins
of the ATX20/24 connector, two (unused) pins were reduced in
size, which creates a DRC violation. While annulus size of pins
is rarely a problem - all pins are created from footprints -,
the DRC check for minimum annulus size was left in place.