Get firmware for the USB adapter started.
For the start, the project CDC-SPI was choosen:
http://www.recursion.jp/avrcdc/cdc-spi.html

This in turn includes V-USB, version 20090822.
http://www.obdev.at/products/vusb/download-de.html

As CDC-SPI is obviously a one-off project, no further
developments are expected there. So only the files neccessary for
our ATtiny45/85 were included and the remaining support for
other CPUs will be removed for more code cleanlyness.

A few minor changes in Makefile and osccal.h wre applied to
make the new sources compilable.