usb adapter: write down everything needed to restart this.
Here, "this" means the ATtiny45/85 based USB-TTL adapter,
especially it's software.

Unless the move to the Microchip MCP2200 turns out to be a failure,
this part of Generation 7 Electronics can be considered to be
abandoned.

Still, this firmware is recognised as USB (CDC-type) device by
PC operating systems just fine, so someone might pick up for
other purposes.